+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/samaritan-proxy/sash
+
+go 1.13
