+++ service/registry.go
@@ -1,4 +1,4 @@
-package registry
+package service
 
 // Instance represents an instance of service.
 type Instance struct {
@@ -8,35 +8,33 @@ type Instance struct {
 }
 
 // EventType indicates the type of event.
-type EventType uint8
+type RegistryEventType uint8
 
+// The following describes the type of registry event.
 const (
-	// EventServiceDelete indicates an event of service delete.
-	EventServiceDelete = iota + 1
-	// EventServiceAdd indicates an event of service's instance add.
-	EventInstanceAdd
-	// EventInstanceDelete indicates an event of service's instance delete.
-	EventInstanceDelete
+	RegistryEventServiceDelete = iota + 1
+	RegistryEventInstanceAdd
+	RegistryEventInstanceDelete
 )
 
 // Event represents an event of service.
-type Event struct {
-	Type      EventType
+type RegistryEvent struct {
+	Type      RegistryEventType
 	Service   string
 	Instances []*Instance
 }
 
 // Registry represents a service registry.
 type Registry interface {
-	// Event returns a event channel which will publish all changes
-	// of the subscribed services.
-	Event() <-chan Event
-
+	// List returns all registerd services.
+	List() ([]string, error)
 	// Get gets all instances of the specified service.
 	Get(service string) ([]*Instance, error)
+
 	// GetAndSub gets all instances of the specified service, and subscribes
 	// the subsequent changes.
 	GetAndSub(service string) ([]*Instance, error)
-	// Unsubscribe unsubscribes changes of the specified service.
-	Unsubscribe(service string) error
+	// Event returns a event channel which will publish all changes
+	// of the subscribed services.
+	Event() <-chan RegistryEvent
 }
